Updates and cleanup for Ktor benchmarks
