+++ frameworks/Kotlin/ktor/Readme.md
@@ -1,4 +1,7 @@
 See subprojects
 
 * [Ktor](ktor/) Ktor using traditional JDBC using various servers
-* [Ktor-asyncdb](ktor-asyncdb/) Ktor with Netty-based PostgreSQL clients
+* [Ktor-jasync](ktor-asyncdb/) Ktor with Netty-based PostgreSQL clients
+* [Ktor-r2dbc](ktor-r2dbc/) Ktor with R2DBC for reactive database access
+* [Ktor-exposed](ktor-exposed/) Ktor with the Exposed database abstraction library
+* [Ktor-pgclient](ktor-pgclient/) Ktor with Reactive PostgreSQL Client
