+++ frameworks/Kotlin/ktor/benchmark_config.json
@@ -101,7 +101,7 @@
         "query_url": "/query/?queries=",
         "fortune_url": "/fortunes",
         "update_url": "/updates?queries=",
-        "port": 8080,
+        "port": 9090,
         "approach": "Realistic",
         "classification": "Fullstack",
         "database": "Postgres",
@@ -125,7 +125,7 @@
         "update_url": "/updates?queries=",
         "fortune_url": "/fortunes",
 
-        "port": 8080,
+        "port": 9090,
         "approach": "Realistic",
         "classification": "Micro",
         "database": "Postgres",
@@ -145,7 +145,7 @@
         "query_url": "/queries?queries=",
         "update_url": "/updates?queries=",
         "fortune_url": "/fortunes",
-        "port": 8080,
+        "port": 9090,
         "approach": "Realistic",
         "classification": "Micro",
         "database": "postgres",
@@ -164,7 +164,7 @@
       "exposed-dao": {
         "db_url": "/db",
         "fortune_url": "/fortunes",
-        "port": 8080,
+        "port": 9090,
         "approach": "Realistic",
         "classification": "Micro",
         "database": "postgres",
