+++ frameworks/Kotlin/ktor/ktor-asyncdb/build.gradle.kts
@@ -13,11 +13,11 @@ repositories {
 }
 
 application {
-    mainClass.set("MainKt")
+    mainClass = "io.ktor.server.netty.EngineMain"
 }
 
-val ktor_version = "2.3.12"
-val kotlinx_serialization_version = "1.7.3"
+val ktor_version = "3.1.2"
+val kotlinx_serialization_version = "1.8.1"
 
 dependencies {
     implementation("org.jetbrains.kotlinx:kotlinx-serialization-json:$kotlinx_serialization_version")
@@ -26,6 +26,7 @@ dependencies {
     implementation("io.ktor:ktor-server-default-headers:$ktor_version")
     implementation("io.ktor:ktor-server-html-builder:$ktor_version")
     implementation("com.github.jasync-sql:jasync-postgresql:2.2.4")
+    implementation("ch.qos.logback:logback-classic:1.5.12")
 }
 
 java {
