+++ frameworks/Kotlin/ktor/ktor-exposed/app/build.gradle.kts
@@ -9,7 +9,7 @@ repositories {
     mavenCentral()
 }
 
-val ktorVersion = "3.1.1"
+val ktorVersion = "3.1.2"
 val kotlinxSerializationVersion = "1.8.0"
 val exposedVersion = "0.59.0"
