+++ frameworks/Kotlin/ktor/ktor-pgclient/src/main/resources/application.conf
@@ -7,6 +7,6 @@ ktor {
     }
 
     application {
-        modules = [org.jetbrains.ktor.benchmarks.MainKt.main]
+        modules = [MainKt.main]
     }
 }
