+++ frameworks/Kotlin/ktor/ktor-r2dbc/pom.xml
@@ -14,8 +14,8 @@
     <properties>
         <kotlin.version>2.1.20</kotlin.version>
         <kotlin.coroutines.version>1.10.1</kotlin.coroutines.version>
-        <ktor.version>3.1.1</ktor.version>
-        <serialization.version>1.8.0</serialization.version>
+        <ktor.version>3.1.2</ktor.version>
+        <serialization.version>1.8.1</serialization.version>
         <kotlinx.html.version>0.12.0</kotlinx.html.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <logback.version>1.5.12</logback.version>
@@ -40,6 +40,11 @@
         <dependency>
             <groupId>org.jetbrains.kotlinx</groupId>
             <artifactId>kotlinx-serialization-json</artifactId>
+            <version>${serialization.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.jetbrains.kotlinx</groupId>
+            <artifactId>kotlinx-serialization-json-io</artifactId>
             <version>${serialization.version}</version>
         </dependency>
         <dependency>
