--- frameworks/Kotlin/ktor/ktor-r2dbc/src/main/kotlin/org/jetbrains/ktor/benchmarks/models/Fortune.kt
@@ -1,6 +1,0 @@
-package org.jetbrains.ktor.benchmarks.models
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class Fortune(val id: Int, var message: String)
