--- frameworks/Kotlin/ktor/ktor-r2dbc/src/main/kotlin/org/jetbrains/ktor/benchmarks/models/Message.kt
@@ -1,11 +1,0 @@
-package org.jetbrains.ktor.benchmarks.models
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class Message(val message: String)
-
-// Cache common messages to reduce allocations
-object MessageCache {
-    val helloWorld = Message("Hello, world!")
-}
