--- frameworks/Kotlin/ktor/ktor-r2dbc/src/main/kotlin/org/jetbrains/ktor/benchmarks/models/World.kt
@@ -1,6 +1,0 @@
-package org.jetbrains.ktor.benchmarks.models
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class World(val id: Int, var randomNumber: Int)
