+++ frameworks/Kotlin/ktor/ktor/pom.xml
@@ -13,12 +13,12 @@
 
     <properties>
         <kotlin.version>2.1.20</kotlin.version>
-        <ktor.version>3.1.1</ktor.version>
-        <serialization.version>1.8.0</serialization.version>
+        <ktor.version>3.1.2</ktor.version>
+        <serialization.version>1.8.1</serialization.version>
         <kotlinx.html.version>0.12.0</kotlinx.html.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <hikaricp.version>5.1.0</hikaricp.version>
-        <logback.version>1.2.13</logback.version>
+        <logback.version>1.5.12</logback.version>
         <mysql.version>8.0.33</mysql.version>
         <postgresql.version>42.7.4</postgresql.version>
     </properties>
@@ -37,6 +37,11 @@
         <dependency>
             <groupId>org.jetbrains.kotlinx</groupId>
             <artifactId>kotlinx-serialization-json</artifactId>
+            <version>${serialization.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.jetbrains.kotlinx</groupId>
+            <artifactId>kotlinx-serialization-json-io</artifactId>
             <version>${serialization.version}</version>
         </dependency>
         <dependency>
