+++ frameworks/Kotlin/ktor/ktor/src/main/kotlin/org/jetbrains/ktor/benchmarks/Models.kt
@@ -1,0 +1,12 @@
+package org.jetbrains.ktor.benchmarks
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class Message(val message: String)
+
+@Serializable
+data class World(val id: Int, var randomNumber: Int)
+
+@Serializable
+data class Fortune(val id: Int, var message: String)
